chore: roomHeader 컴포넌트 스타일 파일 분리
